Porting almost finished
Nearly done! I just need to finish fire rendering and test everything.

Thanks to BlockStates I've been able to remove a lot of code and image
files, but then again there are all these jsons so the mod might not be
much smaller.

I'm getting json problems because it's looking for the variants that
fire blocks have, which I don't want to include in my blocks. Any
advice? I'd really appreciate it.